added ftp log
